requirements.txt
streamlit
pymongo
python-dotenv
boto3